Conflicts:
	libavcodec/hevc.c
	libavcodec/hevc.h
	libavcodec/hevc_refs.c
	libavcodec/hevcdsp.c
	libavcodec/hevcdsp.h
	libavcodec/hevcdsp_template.c
	libavcodec/x86/hevc_idct.asm
	libavcodec/x86/hevc_mc.asm
	libavcodec/x86/hevcdsp.h
	libavcodec/x86/hevcdsp_init.c

+++ libavcodec/hevcdsp_template.c
@@ -340,7 +340,7 @@ static void FUNC(sao_band_filter_0)(uint8_t *_dst, uint8_t *_src,
     int k, y, x;
     int shift  = BIT_DEPTH - 5;
     int16_t *sao_offset_val = sao->offset_val[c_idx];
-    uint8_t sao_left_class  = sao->band_position[c_idx];
+    int sao_left_class  = sao->band_position[c_idx];
 
     stride_dst /= sizeof(pixel);
     stride_src /= sizeof(pixel);
@@ -370,7 +370,7 @@ static void FUNC(sao_edge_filter)(uint8_t *_dst, uint8_t *_src,
         { {  1, -1 }, { -1, 1 } }, // 135 degree
     };
     int16_t *sao_offset_val = sao->offset_val[c_idx];
-    uint8_t sao_eo_class    = sao->eo_class[c_idx];
+    int sao_eo_class    = sao->eo_class[c_idx];
     pixel *dst = (pixel *)_dst;
     pixel *src = (pixel *)_src;
 
@@ -408,7 +408,7 @@ static void FUNC(sao_edge_filter_0)(uint8_t *_dst, uint8_t *_src,
     pixel *dst = (pixel *)_dst;
     pixel *src = (pixel *)_src;
     int16_t *sao_offset_val = sao->offset_val[c_idx];
-    uint8_t sao_eo_class    = sao->eo_class[c_idx];
+    int sao_eo_class    = sao->eo_class[c_idx];
     int init_x = 0, init_y = 0, width = _width, height = _height;
 
     stride_dst /= sizeof(pixel);
@@ -461,7 +461,7 @@ static void FUNC(sao_edge_filter_1)(uint8_t *_dst, uint8_t *_src,
     pixel *dst = (pixel *)_dst;
     pixel *src = (pixel *)_src;
     int16_t *sao_offset_val = sao->offset_val[c_idx];
-    uint8_t sao_eo_class    = sao->eo_class[c_idx];
+    int sao_eo_class    = sao->eo_class[c_idx];
     int init_x = 0, init_y = 0, width = _width, height = _height;
 
     stride_dst /= sizeof(pixel);
